Crate syntastica_parsers_gitdep

syntastica-parsers-gitdep

Collection of tree-sitter parsers for syntastica, using cargo git dependencies.

See the project overview for more information on all parser collections.

List of included parsers

List of parsers included in the some feature

bash

c

cpp

css

go

html

java

javascript

json

kotlin

lua

python

rust

toml (not supported by this collection)

tsx

typescript

yaml

List of parsers additionally included in the most feature

asm

c_sharp

clojure (not supported by this collection)

cmake

comment (not supported by this collection)

dart (not supported by this collection)

diff

elixir

haskell

jsdoc

json5 (not supported by this collection)

jsonc (not supported by this collection)

luap (not supported by this collection)

make

markdown

markdown_inline

nix

php

php_only

printf

regex

ruby

scala

scss (not supported by this collection)

sql

swift (not supported by this collection)

typst (not supported by this collection)

zig

List of parsers additionally included in the all feature

bibtex (not supported by this collection)

dockerfile (not supported by this collection)

ebnf (not supported by this collection)

ejs

erb

fish

gleam

hexdump (not supported by this collection)

julia

lalrpop

latex (not supported by this collection)

llvm (not supported by this collection)

ocaml

ocaml_interface

ql

rush

ursa (not supported by this collection)

verilog (not supported by this collection)

wat (not supported by this collection)

Features

Every supported language has a feature with the same name as the respective public function. Additionally the three feature groups some, most, and all are available.

some — Include parsers for the most widely known supported languages.

most — Implies some. Include parsers for most common languages.

all — Implies most. Include parsers for all supported languages.

docs — Meant to be enabled when building docs

Structs

LanguageSetImpl

An implementation of LanguageSet including all languages in the enabled feature set.

Enums

Lang

An enum of every supported language in the current feature set.

Constants

LANGUAGES

A list of all languages supported by the current feature set.

LANGUAGE_NAMES

A list of all language names supported by the current feature set.

Functions

asmmost or asm

Get the parser for asm.

bashsome or bash

Get the parser for bash.

bibtexall or bibtex

This parser is not supported by this parser collection and thus this function will panic!

csome or c

Get the parser for c.

c_sharpmost or c_sharp

Get the parser for c_sharp.

clojuremost or clojure

This parser is not supported by this parser collection and thus this function will panic!

cmakemost or cmake

Get the parser for cmake.

commentmost or comment

This parser is not supported by this parser collection and thus this function will panic!

cppsome or cpp

Get the parser for cpp.

csssome or css

Get the parser for css.

dartmost or dart

This parser is not supported by this parser collection and thus this function will panic!

diffmost or diff

Get the parser for diff.

dockerfileall or dockerfile

This parser is not supported by this parser collection and thus this function will panic!

ebnfall or ebnf

This parser is not supported by this parser collection and thus this function will panic!

ejsall or ejs

Get the parser for ejs.

elixirmost or elixir

Get the parser for elixir.

erball or erb

Get the parser for erb.

fishall or fish

Get the parser for fish.

gleamall or gleam

Get the parser for gleam.

gosome or go

Get the parser for go.

haskellmost or haskell

Get the parser for haskell.

hexdumpall or hexdump

This parser is not supported by this parser collection and thus this function will panic!

htmlsome or html

Get the parser for html.

javasome or java

Get the parser for java.

javascriptsome or javascript

Get the parser for javascript.

jsdocmost or jsdoc

Get the parser for jsdoc.

jsonsome or json

Get the parser for json.

json5most or json5

This parser is not supported by this parser collection and thus this function will panic!

jsoncmost or jsonc

This parser is not supported by this parser collection and thus this function will panic!

juliaall or julia

Get the parser for julia.

kotlinsome or kotlin

Get the parser for kotlin.

lalrpopall or lalrpop

Get the parser for lalrpop.

latexall or latex

This parser is not supported by this parser collection and thus this function will panic!

llvmall or llvm

This parser is not supported by this parser collection and thus this function will panic!

luasome or lua

Get the parser for lua.

luapmost or luap

This parser is not supported by this parser collection and thus this function will panic!

makemost or make

Get the parser for make.

markdownmost or markdown

Get the parser for markdown.

markdown_inlinemost or markdown_inline

Get the parser for markdown_inline.

nixmost or nix

Get the parser for nix.

ocamlall or ocaml

Get the parser for ocaml.

ocaml_interfaceall or ocaml_interface

Get the parser for ocaml_interface.

phpmost or php

Get the parser for php.

php_onlymost or php_only

Get the parser for php_only.

printfmost or printf

Get the parser for printf.

pythonsome or python

Get the parser for python.

qlall or ql

Get the parser for ql.

regexmost or regex

Get the parser for regex.

rubymost or ruby

Get the parser for ruby.

rushall or rush

Get the parser for rush.

rustsome or rust

Get the parser for rust.

scalamost or scala

Get the parser for scala.

scssmost or scss

This parser is not supported by this parser collection and thus this function will panic!

sqlmost or sql

Get the parser for sql.

swiftmost or swift

This parser is not supported by this parser collection and thus this function will panic!

tomlsome or toml

This parser is not supported by this parser collection and thus this function will panic!

tsxsome or tsx

Get the parser for tsx.

typescriptsome or typescript

Get the parser for typescript.

typstmost or typst

This parser is not supported by this parser collection and thus this function will panic!

ursaall or ursa

This parser is not supported by this parser collection and thus this function will panic!

verilogall or verilog

This parser is not supported by this parser collection and thus this function will panic!

watall or wat

This parser is not supported by this parser collection and thus this function will panic!

yamlsome or yaml

Get the parser for yaml.

zigmost or zig

Get the parser for zig.