Crate syntastica_parsers_gitdep

syntastica-parsers-gitdep

Collection of tree-sitter parsers for syntastica, using cargo git dependencies.

See the project overview for more information on all parser collections.

List of included parsers

List of parsers included in the some feature

bash

c

cpp

css

go

html (not supported by this collection)

java

javascript

json

lua

python

rust

toml

tsx (not supported by this collection)

typescript (not supported by this collection)

yaml

List of parsers additionally included in the most feature

asm

c_sharp

comment

dart (not supported by this collection)

diff

haskell

jsdoc (not supported by this collection)

json5

jsonc (not supported by this collection)

latex

markdown

markdown_inline

php

regex

ruby

scala

scss

List of parsers additionally included in the all feature

ebnf

ejs

erb

hexdump

julia (not supported by this collection)

llvm

ocaml

ocaml_interface

ql

rush

ursa

verilog

wat (not supported by this collection)

Features

Every supported language has a feature with the same name as the respective public function. Additionally the three feature groups some, most, and all are available.

some — Include parsers for the most widely known supported languages.

most — Implies some. Include parsers for most common languages.

all — Implies most. Include parsers for all supported languages.

docs — Meant to be enabled when building docs

Structs

LanguageSetImpl

An implementation of LanguageSet including all languages in the enabled feature set.

Enums

Lang

An enum of every supported language in the current feature set.

Constants

LANGUAGES

A list of all languages supported by the current feature set.

LANGUAGE_NAMES

A list of all language names supported by the current feature set.

Functions

asmmost or asm

Get the parser for asm.

bashsome or bash

Get the parser for bash.

csome or c

Get the parser for c.

c_sharpmost or c_sharp

Get the parser for c_sharp.

commentmost or comment

Get the parser for comment.

cppsome or cpp

Get the parser for cpp.

csssome or css

Get the parser for css.

dartmost or dart

This parser is not supported by this parser collection and thus this function will panic!

diffmost or diff

Get the parser for diff.

ebnfall or ebnf

Get the parser for ebnf.

ejsall or ejs

Get the parser for ejs.

erball or erb

Get the parser for erb.

gosome or go

Get the parser for go.

haskellmost or haskell

Get the parser for haskell.

hexdumpall or hexdump

Get the parser for hexdump.

htmlsome or html

This parser is not supported by this parser collection and thus this function will panic!

javasome or java

Get the parser for java.

javascriptsome or javascript

Get the parser for javascript.

jsdocmost or jsdoc

This parser is not supported by this parser collection and thus this function will panic!

jsonsome or json

Get the parser for json.

json5most or json5

Get the parser for json5.

jsoncmost or jsonc

This parser is not supported by this parser collection and thus this function will panic!

juliaall or julia

This parser is not supported by this parser collection and thus this function will panic!

latexmost or latex

Get the parser for latex.

llvmall or llvm

Get the parser for llvm.

luasome or lua

Get the parser for lua.

markdownmost or markdown

Get the parser for markdown.

markdown_inlinemost or markdown_inline

Get the parser for markdown_inline.

ocamlall or ocaml

Get the parser for ocaml.

ocaml_interfaceall or ocaml_interface

Get the parser for ocaml_interface.

phpmost or php

Get the parser for php.

pythonsome or python

Get the parser for python.

qlall or ql

Get the parser for ql.

regexmost or regex

Get the parser for regex.

rubymost or ruby

Get the parser for ruby.

rushall or rush

Get the parser for rush.

rustsome or rust

Get the parser for rust.

scalamost or scala

Get the parser for scala.

scssmost or scss

Get the parser for scss.

tomlsome or toml

Get the parser for toml.

tsxsome or tsx

This parser is not supported by this parser collection and thus this function will panic!

typescriptsome or typescript

This parser is not supported by this parser collection and thus this function will panic!

ursaall or ursa

Get the parser for ursa.

verilogall or verilog

Get the parser for verilog.

watall or wat

This parser is not supported by this parser collection and thus this function will panic!

yamlsome or yaml

Get the parser for yaml.